+++ src/main/resources/assets/computercraft/lua/rom/apis/gps.lua
@@ -1,4 +1,4 @@
-local expect = _G["~expect"]
+local expect = dofile("rom/modules/main/craftos/expect.lua")
 
 CHANNEL_GPS = 65534
