+++ src/main/resources/assets/computercraft/lua/rom/apis/keys.lua
@@ -1,7 +1,7 @@
 -- Minecraft key code bindings
 -- See http://www.minecraftwiki.net/wiki/Key_codes for more info
 
-local expect = _G["~expect"]
+local expect = dofile("rom/modules/main/craftos/expect.lua")
 
 local tKeys = {
     nil,	 	"one", 		"two", 		"three", 	"four",			-- 1
