+++ src/main/resources/assets/computercraft/lua/rom/apis/settings.lua
@@ -1,4 +1,4 @@
-local expect = _G["~expect"]
+local expect = dofile("rom/modules/main/craftos/expect.lua")
 
 local tSettings = {}
