+++ src/main/resources/assets/computercraft/lua/rom/apis/textutils.lua
@@ -1,4 +1,4 @@
-local expect = _G["~expect"]
+local expect = dofile("rom/modules/main/craftos/expect.lua")
 
 function slowWrite( sText, nRate )
     expect(2, nRate, "number", "nil")
