+++ src/main/resources/assets/computercraft/lua/rom/apis/window.lua
@@ -1,4 +1,4 @@
-local expect = _G["~expect"]
+local expect = dofile("rom/modules/main/craftos/expect.lua")
 
 local tHex = {
     [ colors.white ] = "0",
