+++ src/main/resources/assets/computercraft/lua/rom/programs/advanced/multishell.lua
@@ -1,4 +1,4 @@
-local expect = _G["~expect"]
+local expect = dofile("rom/modules/main/craftos/expect.lua")
 
 -- Setup process switching
 local parentTerm = term.current()
