+++ src/main/resources/assets/computercraft/lua/rom/programs/shell.lua
@@ -1,4 +1,4 @@
-local expect = _G["~expect"]
+local expect = dofile("rom/modules/main/craftos/expect.lua")
 
 local multishell = multishell
 local parentShell = shell
