+++ src/test/resources/test-rom/spec/base_spec.lua
@@ -1,36 +1,4 @@
 describe("The Lua base library", function()
-    describe("expect", function()
-        local e = _G["~expect"]
-
-        it("checks a single type", function()
-            expect(e(1, "test", "string")):eq(true)
-            expect(e(1, 2, "number")):eq(true)
-
-            expect.error(e, 1, nil, "string"):eq("bad argument #1 (expected string, got nil)")
-            expect.error(e, 2, 1, "nil"):eq("bad argument #2 (expected nil, got number)")
-        end)
-
-        it("checks multiple types", function()
-            expect(e(1, "test", "string", "number")):eq(true)
-            expect(e(1, 2, "string", "number")):eq(true)
-
-            expect.error(e, 1, nil, "string", "number"):eq("bad argument #1 (expected string or number, got nil)")
-            expect.error(e, 2, false, "string", "table", "number", "nil")
-                  :eq("bad argument #2 (expected string, table or number, got boolean)")
-        end)
-
-        it("includes the function name", function()
-            local function worker()
-                expect(e(1, nil, "string")):eq(true)
-            end
-            local function trampoline()
-                worker()
-            end
-
-            expect.error(trampoline):eq("base_spec.lua:27: bad argument #1 to 'worker' (expected string, got nil)")
-        end)
-    end)
-
     describe("sleep", function()
         it("validates arguments", function()
             sleep(0)
